requirements.txt
Flask==2.3.3
requests==2.31.0
google-cloud-storage==2.10.0
google-auth==2.23.4
Werkzeug==2.3.7
gunicorn==21.2.0